requirements.txt
fastapi==0.95.2
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.5.0
scikit_image==0.19.3
torch==2.0.1
torchvision==0.15.2
uvicorn==0.22.0
